Move layers, auto resize draw & erase batches
- Added layer moving
- Bucket and flood filling working again
- Added empty cached layer
- Added layer.createBatchAt
- Draw and erase batches get auto resized by their matrix data after
mouseup -> since batch resize jump factor is now 64 by default
- Updated to v0.4.8

+++ src/batch/index.js
@@ -106,7 +106,7 @@ Batch.prototype.kill = function() {
       }
     };
   };
-  if (count <= 0) {
+  if (count <= 0 && !this.isMover) {
     throw new Error(`Failed to kill batch:${this.id}`);
   }
 };
